completed save crud operation
there is a huge mistake done in myself which is i forgot to add the annotation @RequstBody in the conroller save method, the problem is error is give in there is no ids, so its complicated to identify my error. when i send data in the request body, but i did not mention to take requet body in save method, thats why error came. when the save method call, there is nothing to body, cause i did not mention in the save method.