update crud operation implemented
almost same as save operation, but there few important things to mention, when query string should be sent the reg no and, and it take in the method by @PathVariable, and patch mapping used value="/{reg"} to take the reg no what we need to update and then firstly we should set the value for sampleDTO using setReg method.

another important thing is when we send the updated JSON, it should contain the reg field too. if not, update is not done.