get service implemented
use findbyid method in repository layer to find the sample and return it as sampleDTO and there is the get method in api layer and it needs to return bookDTO too as response.

in the first time get method is execute but not get the result, because i forgot to return the dto from api layer. it is important to note that when the result shoud need to return, it should give proper return as JSON.